ISMwithLSTM.SetUpLSTM — Method

function SetUpLSTM(in_size::Int64,C_size::Int64)

Sets up an LSTM neural network (for reference check out this article) where the input dimension in each recurrent step is given by in_size and the dimension of the cell state by C_size. The dimension of the state which the acticle calls 'h' is the same as the dimension of the cell state, so the final output is computed by a single dense layer mapping 'h' to a scalar value (i.e. the estimate for the proximity function at the given time).

ISMwithLSTM.trainLSTM — Function

function trainLSTM(pr::ProxFct,
    paths_to_data::Vector{String},
    train_period::Tuple{Int64, Int64}=(1948,1980),
    test_period::Tuple{Int64, Int64}=(1981,2010);
    C_dim::Int64 = 5,
    Tr::Int64 = 365+365,
    epochs::Int64 = 100,
    λ1::Float64 = 0.,
    λ2::Float64 = 0.,
    learning_rate::Float64 = 1e-2,
    reduce_learning_rate::Int64 = 20)

Trains a given LSTM network on training data and hyperparameters specified by the arguments of the function

Arguments

pr::ProxFct: Instance of ProxFct holding the information about the proximity function

paths_to_data::Vector{String}: Array of strings describing the paths to the data which should be used for training

train_period::Tuple{Int64, Int64}=(1948,1980): Start and end year of training

test_period::Tuple{Int64, Int64}=(1981,2010): Start and end year of testing

C_dim::Int64 = 5: Dimension of the cell state of the LSTM (for reference check out this article)

Tr::Int64 = 366+365: Transient time in days. The model will not be trained on the first Tr number of days of the training set

epochs::Int64 = 100: Number of maximum epochs (early stopping possible if loss diverges or loss reaches a platteu)

λ1::Float64 = 0.: Weight of the L1 regularisation

λ2::Float64 = 0.: Weight of the L2 regularisation

learning_rate::Float64 = 1e-3: Initial learning rate. It reduces by a factor of two every 30 epochs.

reduce_learning_rate::Int64 = 20: Reduces learing rate by a factor of two in regular steps.

Returns

Trained LSTM

Array of train losses (one loss value per epoch)

Array of test losses (one loss value per epoch)

Loss function: takes arguments (data,prox,LSTM)

ISMwithLSTM.saveLSTM — Method

function saveLSTM(LSTM,path::String)

Saves the LSTM as a '.bson' file at the given location.

Arguments

LSTM: LSTM one wants to save

path::String: e.g. "parameters/my_lstm.bson"

ISMwithLSTM.loadLSTM — Method

function loadLSTM(path::String)

Returns the LSTM saved at the given location.

Arguments

path::String: e.g. "parameters/my_lstm.bson"

ISMwithLSTM.saveFct — Method

function saveFct(loss,path::String)

Saves the loss function as a '.bson' file at the given location.

Arguments

loss: loss function one wants to save

path::String: e.g. "losses/my_loss.bson"

ISMwithLSTM.loadFct — Method

function loadFct(path::String)

Returns the Loss function saved at the given location.

Arguments

path::String: e.g. "losses/my_loss.bson"

ISMwithLSTM.ProxFct — Method

function ProxFct(path_to_data::String)

Returns an instance of the struct ProxFct with fields filled with data according to the given path. This struct stores relevant data for the proximity function. It is overloaded with a function returning the data of a respective intervall of years.

Fields

ProxData::Matrix{Real}: Matrix with first column filled with the years and second column filled with the vaules of ΔTT

onsets::Matrix{Real}: First column with years and second column with the onset date of the respective year

Examples

julia> pr = ProxFct("datafile.txt")
ProxFct(Real[1948.0 0.0; 1948.0 0.0; … ; 2022.0 0.0; 2022.0 0.0], Real[1948 157; 1949 151; … ; 2021 148; 2022 145])

julia> pr((2010,2020))
(Real[0.350210970464135, 0.3544303797468354, 0.3586497890295358, 0.36286919831223624, 0.36708860759493667, 0.3713080168776371, 
0.3755274261603375, 0.37974683544303794, 0.38396624472573837, 0.3881856540084388  …  0.3114035087719298, 0.3157894736842105, 
0.3201754385964912, 0.32456140350877194, 0.3289473684210526, 0.3333333333333333, 0.33771929824561403, 0.3421052631578947, 
0.3464912280701754, 0.3508771929824561], Real[155, 159, 152, 147, 161, 158, 161, 150, 151, 160, 156])