ISMwithLSTM.RunLSTM — Method

function RunLSTM(LSTM, paths_to_data::Vector{String},end_year::Int)

Runs given LSTM on the data over the given period

Arguments

LSTM

paths_to_data::Vector{String}: Array of strings describing the paths to the data which should be used for training

end_year::Int: Year until which the LSTM is run

ISMwithLSTM.CalculateLoss — Method

function CalculateLoss(loss, LSTM, paths_to_data::Vector{String},pr::ProxFct,run_period::Tuple{Int64,Int64})

Calculates loss over the given run_period.

Arguments

loss: The loss function which is used

LSTM

paths_to_data::Vector{String}

pr::ProxFct

run_period::Tuple{Int64,Int64}

Returns

loss_value::Float64

ISMwithLSTM.OnsetDayPrediction — Method

function OnsetDayPrediction(LSTM, paths_to_data::Vector{String}, yr::Int; t_1::Int = 60, t_2::Int = 70)

Arguments

LSTM

paths_to_data::Vector{String}

yr::Int: year to predict the onset for

Keyword Arguments

t_1::Int = 60: corresponds to the number of days before January 1st of the prediction year

t_2::Int = 70: correpsonds to the number of days after January 1st of the prediction year

Returns

ISM Onset Day

ISMwithLSTM.PlotProximity — Method

function PlotProximity(LSTM, pr::ProxFct ,paths_to_data::Vector{String}, time_period::Tuple{Int64,Int64},save_directory::String)

Plots the real Proximity Function vs. the learned one on the given time_period

Arguments

LSTM-

pr::ProxFct

paths_to_data::Vector{String}

time_period::Tuple{Int64,Int64}

save_directory::String

ISMwithLSTM.PlotOnsetComparsion — Method

function PlotOnsetComparsion(LSTM, 
    paths_to_data::Vector{String}, 
    pr::ProxFct,
    time_period::Tuple{Int64,Int64};
    t_1= 60::Int, 
    t_2=70::Int, 
    save_directory::String)

Plots the predicted onsets vs the actual ones.

Arguments

LSTM

paths_to_data::Vector{String}

pr::ProxFct

time_period::Tuple{Int64,Int64}

Keyword Arguments

t_1= 60::Int

t_2=70::Int

save_directory::String

ylim::Tuple{Int64,Int64}

ISMwithLSTM.WriteOutLeadTimeAnalysis — Method

function WriteOutLeadTimeAnalysis(LSTM, 
    paths_to_data::Vector{String}, 
    pr::ProxFct,
    save_directory::String; 
    lead_time_range = (60,110),
    time_period = (1981,2020)::Tuple{Int64,Int64})

Writes out RSME and Correlation of the onset days for the given lead-time range to a .txt file.

Arguments

LSTM

paths_to_data::Vector{String}

pr::ProxFct

save_directory::String

Keyword Arguments

lead_time_range = (60,110)

time_period = (1981,2020)::Tuple{Int64,Int64})

Returns

vector of standard deviation of the predicted onset days

vector of correlation between predicted and actual onset days

ISMwithLSTM.PlotLeadTimeAnalysis — Method

function PlotLeadTimeAnalysis( 
    path_to_data::String, 
    save_directory::String; 
    lead_time_range = (60,110),
    time_period = (1981,2020)::Tuple{Int64,Int64})

Plots the RSME and Correlation of the onset days for the given lead-time range. Takes in data from the function WriteOutLeadTimeAnalysis

Arguments

path_to_data::String

save_directory::String

Keyword Arguments

lead_time_range = (60,110)

time_period = (1981,2020)::Tuple{Int64,Int64})